separated  functions for testing
